use baseUrl from gatsby-config, upgrade depdencies
