+++ package.json
@@ -5,25 +5,25 @@
   "author": "Code For Muenster <codeformuenster@gmail.com>",
   "dependencies": {
     "@turf/turf": "^5.1.6",
-    "deck.gl": "^6.3.1",
-    "gatsby": "^2.0.0",
-    "gatsby-plugin-manifest": "^2.0.2",
-    "gatsby-plugin-netlify": "^2.0.0",
-    "gatsby-plugin-offline": "^2.0.5",
-    "gatsby-plugin-react-helmet": "^3.0.0",
-    "gatsby-plugin-typography": "^2.2.1",
-    "got": "^9.2.2",
+    "deck.gl": "^6.3.2",
+    "gatsby": "^2.0.72",
+    "gatsby-plugin-manifest": "^2.0.12",
+    "gatsby-plugin-netlify": "^2.0.6",
+    "gatsby-plugin-offline": "^2.0.20",
+    "gatsby-plugin-react-helmet": "^3.0.4",
+    "gatsby-plugin-typography": "^2.2.2",
+    "got": "^9.5.0",
     "kinto": "^12.0.2",
     "lodash.camelcase": "^4.3.0",
-    "mapbox-gl": "0.51.0",
+    "mapbox-gl": "0.52.0-beta.2",
     "normalize.css": "^8.0.0",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
     "react-helmet": "^5.2.0",
-    "react-map-gl": "^4.0.2",
-    "react-redux": "^5.0.7",
-    "react-typography": "^0.16.13",
-    "typography": "^0.16.17"
+    "react-map-gl": "^4.0.5",
+    "react-redux": "^6.0.0",
+    "react-typography": "^0.16.18",
+    "typography": "^0.16.18"
   },
   "license": "Apache-2.0",
   "scripts": {
@@ -37,7 +37,7 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-react": "^7.11.1",
     "gatsby-plugin-eslint": "^2.0.1",
-    "prettier": "^1.14.2"
+    "prettier": "^1.15.3"
   },
   "repository": {
     "type": "git",
