+++ src/containers/Map/Map.js
@@ -1,17 +1,48 @@
-import Map from '../../components/Map/Map';
-import { connect } from 'react-redux';
+import React from 'react'
+import Map from '../../components/Map/Map'
+import { connect } from 'react-redux'
+import { graphql, StaticQuery } from 'gatsby'
+import Client from '../../helpers/client'
 
-export function mapStateToProps({ allMarkets, marketData }) {
-  return { allMarkets, marketData }
-}
+const StaticQueryMap = props => (
+  <StaticQuery
+    query={graphql`
+      {
+        site {
+          siteMetadata {
+            baseUrl
+          }
+        }
+      }
+    `}
+    render={data => (
+      <Map {...props} client={Client(data.site.siteMetadata.baseUrl)} />
+    )}
+  />
+)
 
-export function mapDispatchToProps(dispatch) {
+const mapStateToProps = ({ allBooths, allMarkets, viewport, filterData }) => ({
+  allBooths,
+  allMarkets,
+  viewport,
+  filterData,
+})
+const mapDispatchToProps = dispatch => {
   return {
-    setAllMarkets: allMarkets =>
-      dispatch({ type: `SET_ALL_MARKETS`, payload: allMarkets }),
+    setAllBooths: markets =>
+      dispatch({ type: 'SET_ALL_BOOTHS', payload: markets }),
+    setAllMarkets: markets =>
+      dispatch({ type: 'SET_ALL_MARKETS', payload: markets }),
     setSelectedMarket: market =>
-      dispatch({ type: `SET_SELECTED_MARKET`, payload: market })
+      dispatch({ type: `SET_SELECTED_MARKET`, payload: market }),
+    setSelectedBooth: booth =>
+      dispatch({ type: `SET_SELECTED_BOOTH`, payload: booth }),
+    setViewport: viewport =>
+      dispatch({ type: `SET_VIEWPORT`, payload: viewport }),
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(Map);
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(StaticQueryMap)
